[rename] #9 폴더명 변경
